Add OCSF common module with CWE, CVSS, and CVE classes and corresponding tests
